# Frontend/src/Adapters/seatpicker.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Infrastructure",
    "version": "1.0"
  },
  "paths": {
...
    "/reservation": {
      "post": {
        "tags": [
          "reservation"
        ],
        "operationId": "Seatpicker.Infrastructure.Entrypoints.Http.Reservation.Create.Endpoint (Infrastructure)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReservationCreateRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReservationCreateRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReservationCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/lan": {
      "post": {
        "tags": [
          "lan"
        ],
        "operationId": "Seatpicker.Infrastructure.Entrypoints.Http.Lan.Create.Endpoint (Infrastructure)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LanCreateRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LanCreateRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LanCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/discord/renew": {
      "post": {
        "tags": [
          "discord"
        ],
        "operationId": "Seatpicker.Infrastructure.Authentication.Discord.DiscordAuthenticationController.Renew (Infrastructure)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenewRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RenewRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RenewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/discord/login": {
      "post": {
        "tags": [
          "discord"
        ],
        "operationId": "Seatpicker.Infrastructure.Authentication.Discord.DiscordAuthenticationController.Login (Infrastructure)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/discord/roles": {
      "get": {
        "tags": [
          "discord"
        ],
        "operationId": "Seatpicker.Infrastructure.Authentication.Discord.DiscordAuthenticationController.GetRoles (Infrastructure)",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DiscordRoleMappingResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DiscordRoleMappingResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DiscordRoleMappingResponse"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "discord"
        ],
        "operationId": "Seatpicker.Infrastructure.Authentication.Discord.DiscordAuthenticationController.PutRoles (Infrastructure)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiscordRoleMappingRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DiscordRoleMappingRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DiscordRoleMappingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/lan/{id}": {
      "get": {
        "tags": [
          "lan"
        ],
        "operationId": "Seatpicker.Infrastructure.Entrypoints.Http.Lan.Get.Endpoint (Infrastructure)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LanGetResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LanGetResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LanGetResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "lan"
        ],
        "operationId": "Seatpicker.Infrastructure.Entrypoints.Http.Lan.Update.Endpoint (Infrastructure)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LanUpdateRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LanUpdateRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LanUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/reservation/{id}": {
      "put": {
        "tags": [
          "reservation"
        ],
        "operationId": "Seatpicker.Infrastructure.Entrypoints.Http.Reservation.Move.Endpoint (Infrastructure)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReservationMoveRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReservationMoveRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReservationMoveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "reservation"
        ],
        "operationId": "Seatpicker.Infrastructure.Entrypoints.Http.Reservation.Remove.Endpoint (Infrastructure)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DiscordRoleMapping": {
        "type": "object",
        "properties": {
          "discordRoleId": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        },
        "additionalProperties": false
      },
      "DiscordRoleMappingRequest": {
        "type": "object",
        "properties": {
          "mappings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DiscordRoleMapping"
            }
          }
        },
        "additionalProperties": false
      },
      "DiscordRoleMappingResponse": {
        "type": "object",
        "properties": {
          "discordRoleId": {
            "type": "string"
          },
          "discordRoleName": {
            "type": "string"
          },
          "discordRoleColor": {
            "type": "integer",
            "format": "int32"
          },
          "discordRoleIcon": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        },
        "additionalProperties": false
      },
...
      "LanCreateRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "background": {
            "type": "string",
            "format": "byte"
          }
        },
        "additionalProperties": false
      },
      "LanGetResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "background": {
            "type": "string",
            "format": "byte"
          }
        },
        "additionalProperties": false
      },
      "LanUpdateRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "background": {
            "type": "string",
            "format": "byte",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "RenewRequest": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "ReservationCreateRequest": {
        "type": "object",
        "properties": {
          "seatId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
...
      "ReservationMoveRequest": {
        "type": "object",
        "properties": {
          "fromSeatId": {
            "type": "string",
            "format": "uuid"
          },
          "toSeatId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "Role": {
        "enum": [
          "Admin",
          "Operator",
          "User"
        ],
        "type": "string"
      },
...
      "TokenResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "nick": {
            "type": "string"
          },
          "avatar": {
            "type": "string",
            "nullable": true
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Role"
            }
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Bearer token acquired through the Discord authentication endpoints",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}